Move new API into a separate file for now
